requirements.txt
scikit-learn
pandas
numpy
mlflow
pyyaml
joblib
fastapi
uvicorn
python-multipart